doc: document the direction field in listpeerchannels.
We mentioned it in the schema (so it was allowed), but we didn't document it!

+++ cln-rpc/src/model.rs
@@ -7482,7 +7482,7 @@ pub mod responses {
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub closer: Option<ChannelSide>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub direction: Option<u32>,
+	    pub direction: Option<i64>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub dust_limit_msat: Option<Amount>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
